Update add Option
Add an options Advanced Gcode Comments
If enabled, this option allows more comments to be added to the modified GCode (Existing case).
The use of more comments allows better control and understanding of the modifications made, but increases the size of the final code and sometimes too many comments can produce Issue on the printer.